[MONDRIAN-2615] Type of level member properties is passed as null - ignores schema

+++ mondrian/src/main/java/mondrian/spi/impl/OracleDialect.java
@@ -182,10 +182,8 @@ public class OracleDialect extends JdbcDialectImpl {
             } else if (scale == -127 && precision ==0) {
                 type = SqlStatement.Type.INT;
             } else if (scale == 0 && (precision == 38 || precision == 0)) {
-                // NUMBER(38, 0) is conventionally used in
-                // Oracle for integers of unspecified precision, so let's be
-                // bold and assume that they can fit into an int.
-                type = SqlStatement.Type.INT;
+                // NUMBER(38, 0) is conventionally used in Oracle for integers of unspecified precision
+                type = SqlStatement.Type.LONG;
             } else if (scale == 0 && precision <= 9) {
                 // An int (up to 2^31 = 2.1B) can hold any NUMBER(10, 0) value
                 // (up to 10^9 = 1B).
